refactor(#775): split oversized modules by responsibility
- combat_system.py (650→440 LOC): extract AoE/explosion handlers into
  combat_system_explosions.py via _CombatSystemExplosionsMixin; pattern
  mirrors the existing shared combat_explosions.py / combat_visuals.py split

- raycaster.py (893→580 LOC): extract wall-column rendering dispatch and
  sprite context helpers into raycaster_wall_dispatch.py via
  _RaycasterWallDispatch mixin; _PerRayLists dataclass moved to mixin module
  and re-exported from raycaster.py for backward compat

- raycaster_rendering.py (671→331 LOC): extract floor/ceiling/sky gradients
  and minimap rendering into raycaster_environment.py; original names
  re-exported via noqa:F401 shims so all existing imports continue to work

All public APIs preserved. ruff + black pass. SPEC.md bumped to 1.1.33.
