+++ SPEC.md
@@ -10,8 +10,8 @@
 | **Primary Language(s)** | Python 3.10+ (Pygame), JavaScript (Three.js for web) |
 | **License**             | MIT                                                  |
 | **Current Version**     | N/A                                                  |
-| **Spec Version**        | 1.1.32                                               |
-| **Last Spec Update**    | 2026-04-14                                           |
+| **Spec Version**        | 1.1.33                                               |
+| **Last Spec Update**    | 2026-04-16                                           |
 
 ## 2. Purpose & Mission
 
@@ -409,6 +409,7 @@ Active development. Core games (F1-F6, F8-F9) fully implemented and tested. F7 (
 
 | Date       | Version | Changes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
 | ---------- | ------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| 2026-04-16 | 1.1.33  | Refactor(#775): Split oversized modules by responsibility — `combat_system.py` (650→440 LOC) extracts AoE/explosion handlers into `combat_system_explosions.py` via `_CombatSystemExplosionsMixin`; `raycaster.py` (893→580 LOC) extracts wall-column dispatch into `raycaster_wall_dispatch.py` via `_RaycasterWallDispatch` mixin; `raycaster_rendering.py` (671→331 LOC) extracts floor/ceiling/sky and minimap rendering into `raycaster_environment.py`. All public APIs preserved; `_PerRayLists` re-exported from `raycaster.py` for backward compat. |
 | 2026-04-14 | 1.1.32  | Refactor(#746): P1 remediation — decomposed oversized functions (>40 LOC) across 6 files (`raycaster.py`, `enemy.py`, `tetris.py`, `analyze_completist_data.py`, `mypy_fixers.py`, `mypy_autofix_agent.py`) into focused helpers ≤30 LOC; added `_PerRayLists` dataclass for zero-overhead per-ray data bundling; added 42 TDD tests for all newly extracted helpers.                                                                                                                                                                           |
 | 2026-04-14 | 1.1.31  | Refactor(#744): split 3 monolithic scripts into focused modules — `mypy_autofix_agent.py` (731→262 LOC) extracts fix strategies and data types into `mypy_fixers.py`; `run_assessment.py` (628→163 LOC) extracts 15 per-category analysis functions into `assessment_modules.py`; `combat_manager.py` (624→231 LOC) extracts explosion helpers into `combat_explosions.py` and shot visual helpers into `combat_visuals.py`. All public APIs preserved.                                                                                         |
 | 2026-04-14 | 1.1.28  | Raised statement coverage toward 80% by adding 8 test files (126+ tests): Zombie Survival atmosphere_manager, weapon_system, screen_event_handler, renderer, and game delegation; Duum renderer and screen_event_handler extensions; Force Field combat_actions. Addresses issue #738.                                                                                                                                                                                                                                                          |
